refactor for organization
