+++ src/main/java/com/accbdd/complicated_bees/compat/emi/recipe/FlowerTypeEmiRecipe.java
@@ -4,8 +4,8 @@ import com.accbdd.complicated_bees.ComplicatedBees;
 import com.accbdd.complicated_bees.bees.Flower;
 import com.accbdd.complicated_bees.bees.GeneticHelper;
 import com.accbdd.complicated_bees.compat.emi.ComplicatedBeesEMI;
-import com.accbdd.complicated_bees.compat.emi.EmiBlock;
-import com.accbdd.complicated_bees.compat.emi.EmiFlowerBlocks;
+import com.accbdd.complicated_bees.compat.emi.ingredient.EmiBlock;
+import com.accbdd.complicated_bees.compat.emi.ingredient.EmiFlowerBlocks;
 import com.accbdd.complicated_bees.registry.FlowerRegistration;
 import dev.emi.emi.api.recipe.EmiRecipe;
 import dev.emi.emi.api.recipe.EmiRecipeCategory;
